rust: Fix remaining clippy warnings
warning: casting to the same type is unnecessary (`usize` -> `usize`)
   --> libs/rust/base/src/tcu.rs:877:31
    |
877 |         let regs = EXT_REGS + UNPRIV_REGS as usize;
    |                               ^^^^^^^^^^^^^^^^^^^^ help: try: `UNPRIV_REGS`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
    = note: `#[warn(clippy::unnecessary_cast)]` on by default

warning: this function has too many arguments (8/7)
   --> libs/rust/resmng/src/resources/tiles.rs:122:5
    |
122 | /     pub fn load_mux<A, M>(
123 | |         &mut self,
124 | |         name: &str,
125 | |         mem_size: usize,
...   |
130 | |         mut get_mod: M,
131 | |     ) -> Result<(), Error>
    | |__________________________^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#too_many_arguments
    = note: `#[warn(clippy::too_many_arguments)]` on by default

+++ src/libs/rust/base/src/tcu.rs
@@ -872,9 +872,9 @@ impl TCU {
     /// Prints the given message into the gem5 log
     pub fn print(s: &[u8]) -> usize {
         #[cfg(any(feature = "hw22", feature = "hw23"))]
-        let regs = EXT_REGS + UNPRIV_REGS + (128 * EP_REGS) as usize;
+        let regs = EXT_REGS + UNPRIV_REGS + (128 * EP_REGS);
         #[cfg(not(any(feature = "hw22", feature = "hw23")))]
-        let regs = EXT_REGS + UNPRIV_REGS as usize;
+        let regs = EXT_REGS + UNPRIV_REGS;
 
         let s = &s[0..cmp::min(s.len(), PRINT_REGS * mem::size_of::<Reg>() - 1)];
